@@ -11,9 +11,9 @@
     "@testing-library/user-event": "^12.1.10",
     "@types/jest": "^26.0.15",
     "@types/node": "^12.0.0",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "antd": "^4.15.4",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
+    "antd": "^4.21.6",
     "babel-eslint": "^10.1.0",
     "babel-jest": "^26.6.0",
     "babel-loader": "8.1.0",
@@ -59,19 +59,19 @@
     "postcss-safe-parser": "5.0.2",
     "prompts": "2.4.0",
     "pygmentize-bundled": "^2.3.0",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-app-polyfill": "^2.0.0",
-    "react-copy-to-clipboard": "^5.0.4",
+    "react-copy-to-clipboard": "^5.1.0",
     "react-dev-utils": "^11.0.3",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-markdown": "^5.0.3",
     "react-markdown-editor-lite": "^1.2.4",
     "react-refresh": "^0.8.3",
-    "react-router": "^5.2.0",
+    "react-router": "^6.3.0",
     "react-router-config": "^5.1.1",
-    "react-router-dom": "^5.2.0",
-    "react-syntax-highlighter": "^15.4.3",
-    "recoil": "^0.2.0",
+    "react-router-dom": "6",
+    "react-syntax-highlighter": "^15.5.0",
+    "recoil": "^0.7.4",
     "remark-gfm": "^1.0.0",
     "resolve": "1.18.1",
     "resolve-url-loader": "^3.1.2",
@@ -119,9 +119,9 @@
     "@types/markdown-it": "^12.0.1",
     "@types/marked": "^3.0.0",
     "@types/react-copy-to-clipboard": "^5.0.1",
-    "@types/react-router-config": "^5.0.2",
+    "@types/react-router-config": "^5.0.6",
     "@types/react-router-dom": "^5.1.7",
-    "@types/react-syntax-highlighter": "^13.5.0",
+    "@types/react-syntax-highlighter": "^15.5.3",
     "@typescript-eslint/eslint-plugin": "^4.22.1",
     "@typescript-eslint/parser": "^4.22.1",
     "colors": "^1.4.0",
